+++ src/bin_snoop/main_snoop.ml
@@ -521,7 +521,7 @@ let () =
   @@ Tezos_base_unix.Internal_event_unix.(
        init
          ~lwt_log_sink:Lwt_log_sink_unix.default_cfg
-         ~configuration:Configuration.default)
+         ~configuration:Configuration.lwt_log)
        ()
 
 let () =